requirements.txt
# Core
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0

# LLM & AI
google-generativeai>=0.3.0

# Vector Embeddings
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4

# Text Processing
emoji>=2.8.0
urlextract>=1.8.0
python-dateutil>=2.8.2

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0
wordcloud>=1.9.2

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0

# Database (optional for caching)
SQLAlchemy>=2.0.0